pyproject.toml
[project]
name = "yassirsimageprocessing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mutagen>=1.47.0",
    "pillow>=12.1.1",
    "soundfile>=0.13.1",
]
